requirements.txt
chromadb>=0.4.18
google.generativeai
pymupdf4llm
pydantic-settings
tqdm